// src/app/common/model/set-part.model.ts
export class SetPart {
  id?: any;
  organizationId?: any;
  remark?: any;
  roleCode?: any;
  roleName?: any;
  idt?: any;
  udt?: any;
}
export class AddSetPart {
  id?: any;
  organizationId?: any;
  remark?: any;
  roleCode?: any;
  roleName?: any;
  idt?: any;
  udt?: any;
}
export class ModifySetPart {
  id?: any;
  organizationId?: any;
  remark?: any;
  roleCode?: any;
  roleName?: any;
  idt?: any;
  udt?: any;
}